supernova_agent/requirements.txt
web3==6.15.1
aiohttp==3.9.3
python-telegram-bot[job-queue]==21.6
python-dotenv==1.0.1
eth-abi==5.0.1
eth-account==0.11.0
google-genai>=1.0.0
